Move Device struct definition to the separate file
